refactor(dialogs): moved dialogs in dedicated module
Moved all python dialog files to a dedicated ui module. Cleaner structure and avoids initalization of
the entrie ui module upon importing a dialog.
All import statements upated accordingly.
