+++ src/ui/__init__.py
@@ -1,5 +1,2 @@
-from .settings_dialog import SettingsDialog
-from .epsg_dialog import EpsgDialog
-from .preset_dialog import PresetDialog
 from . import icons as Icons                # Uppercase to be inline with other imports?
 from .menus import MainMenu
