+++ src/ui/dialogs/__init__.py
@@ -1,0 +1,3 @@
+from .epsg_dialog import EpsgDialog
+from .preset_dialog import PresetDialog
+from .settings_dialog import SettingsDialog
